[Solution] Add methods to Node to evaluate its dirtiness
Furthermore, the 'Node::to_string()' and 'Node::repr()' methods have
been altered to reflect the dirtiness of a node.

IMPORTANT: The 'Node::is_very_dirty()' method is to be implemented in
the future. It should evaluate whether the Node should be cached or not
based on said Node's dirty_counter_.