[Solution] Add caching table
The global caching table maps a node's id to a cache entry. The latter
is a lightweight object which only holds the most necessary data to
recreate the cached node. To achieve this behaviour each Node subclass
must implement the 'Node::cache()' method which creates a suitable
CacheEntry for said node.